Improved NIS support
Added use_nis_maps and nis_master_name parameters to control the
inclusion of a NIS auto.master file

Added parameter 'maptype' to the autofs::map class. This adds the
possibility to additionally specify the maptype of the autofs map to
enable alternative data sources.

NOTE: When specifying the maptype the source reference will be to $key,
instead if /etc/auto.$key when not specifying it. This is to retain
legacy compability.